requirements_dev.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-cors==0.8.1
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
apex==0.1
attrs==25.3.0
black==25.1.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
codetiming==1.4.0
colorful==0.5.7
datasets==4.1.0
dill==0.4.0
distlib==0.4.0
einops==0.8.1
filelock==3.13.1
flash_attn==2.8.1
frozenlist==1.7.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.74.0
hf-xet==1.1.10
huggingface-hub==0.34.5
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.7.0
isort==6.0.1
Jinja2==3.1.4
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Markdown==3.9
MarkupSafe==2.1.5
megatron-core==0.15.0rc3
Megatron-LM-AutoTuner==0.0.1
ml_dtypes==0.5.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.8.55
omegaconf==2.3.0
onnx==1.19.0
onnx-ir==0.1.9
onnxscript==0.3.1
opencensus==0.11.4
opencensus-context==0.1.3
opentelemetry-api==1.37.0
opentelemetry-exporter-prometheus==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
peft==0.17.1
pillow==11.0.0
pip==25.2
platformdirs==4.4.0
prometheus_client==0.22.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.0.0
py-spy==0.4.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==3.0.1
pydantic==2.11.9
pydantic_core==2.33.2
pylatexenc==2.10
python-dateutil==2.9.0.post0
pytz==2025.2
pyvers==0.1.0
PyYAML==6.0.2
ray==2.49.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
sentry-sdk==2.38.0
setuptools==78.1.1
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sympy==1.13.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensordict==0.10.0
tokenizers==0.21.4
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchdata==0.11.0
torchvision==0.22.1+cu128
tqdm==4.67.1
transformer_engine==2.8.0.dev0+f378eaf2
transformers==4.55.4
triton==3.3.1
typing_extensions==4.12.2
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
verl==0.5.0.dev0
virtualenv==20.34.0
wandb==0.21.4
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
